requirements-test.txt
numpy==2.2.2
pillow==11.1.0
pytest==8.3.3
pytest-cov==5.0.0
